moves Flag components to folder
